Make vardbapi._aux_get() return _mtime_ as a long (never as a float).
svn path=/main/trunk/; revision=12331

+++ pym/portage/dbapi/vartree.py
@@ -1176,7 +1176,7 @@ class vardbapi(dbapi):
 		results = []
 		for x in wants:
 			if x == "_mtime_":
-				results.append(st.st_mtime)
+				results.append(long(st.st_mtime))
 				continue
 			try:
 				myf = open(os.path.join(mydir, x), "r")
